Start moving merge impl to cell level
