+++ src/components/Spreadsheet/SpreadsheetCell.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import ScrollerCell from '../Scroller/ScrollerCell';
 import SpreadsheetContext from './SpreadsheetContext';
-import { getMergedCellSize } from './utils';
+import { getMergedCellSize, getMergedCellPosition } from './utils';
 
 /**
  * @param {import('./').SpreadsheetCellProps} props 
@@ -12,51 +12,118 @@ const SpreadsheetCell = ({
   columns,
   rowIndex,
   columnIndex,
+  overscrolled,
   children,
   ...props
 }) => {
 
-  /*const {
+  const {
     defaultRowHeight,
     defaultColumnWidth,
     fixRows,
-    fixColumns,
-    specialRowsCount,
-    specialColumnsCount
+    fixColumns
   } = useContext(SpreadsheetContext);
 
-  const width = getMergedCellSize({
-    count: mergedRange.end.column - mergedRange.start.column,
-    meta: columns,
-    startIndex: columnIndex,
-    defaultSize: defaultColumnWidth
-  });
-  const height = getMergedCellSize({
-    count: mergedRange.end.row - mergedRange.start.row,
-    meta: rows,
-    startIndex: rowIndex,
-    defaultSize: defaultRowHeight
-  });
+  let element;
 
-  const fixWidth = fixColumns && columnIndex <= fixColumns && getMergedCellSize({
-    count: fixColumns - mergedRange.start.column,
-    meta: columns,
-    startIndex: columnIndex,
-    defaultSize: defaultColumnWidth
-  });
+  if (mergedRange) {
 
-  const fixHeight = fixRows && rowIndex <= fixRows && getMergedCellSize({
-    count: fixRows - mergedRange.start.row,
-    meta: rows,
-    startIndex: rowIndex,
-    defaultSize: defaultRowHeight
-  });*/
+    const columnIndex = mergedRange.start.column;
+    const rowIndex = mergedRange.start.row;
 
-  return (
-    <ScrollerCell {...props}>
-      {children}
-    </ScrollerCell>
-  );
+    const isFixedColumnArea = columnIndex <= fixColumns;
+    const isFixedRowArea = rowIndex <= fixRows;
+    
+    const top = getMergedCellPosition({
+      meta: rows,
+      index: rowIndex,
+      defaultSize: defaultRowHeight
+    });
+    const left = getMergedCellPosition({
+      meta: columns,
+      index: columnIndex,
+      defaultSize: defaultColumnWidth
+    });
+
+    const width = getMergedCellSize({
+      count: mergedRange.end.column - mergedRange.start.column,
+      meta: columns,
+      startIndex: columnIndex,
+      defaultSize: defaultColumnWidth
+    });
+    const height = getMergedCellSize({
+      count: mergedRange.end.row - mergedRange.start.row,
+      meta: rows,
+      startIndex: rowIndex,
+      defaultSize: defaultRowHeight
+    });
+
+    const fixWidth = fixColumns && columnIndex <= fixColumns && getMergedCellSize({
+      count: fixColumns - mergedRange.start.column,
+      meta: columns,
+      startIndex: columnIndex,
+      defaultSize: defaultColumnWidth
+    });
+
+    const fixHeight = fixRows && rowIndex <= fixRows && getMergedCellSize({
+      count: fixRows - mergedRange.start.row,
+      meta: rows,
+      startIndex: rowIndex,
+      defaultSize: defaultRowHeight
+    });
+
+    /*if (isFixedRowArea && isFixedColumnArea) {
+      element = (
+        <ScrollerCell {...props}>
+          {children}
+        </ScrollerCell>
+      );
+    }
+    if (isFixedColumnArea) {
+      element = (
+        <ScrollerCell {...props}>
+          {children}
+        </ScrollerCell>
+      );
+    }
+    if (isFixedRowArea) {
+      element = (
+        <ScrollerCell {...props}>
+          {children}
+        </ScrollerCell>
+      );
+    }*/
+    // Value level
+    const style = {
+      position: 'absolute',
+      top,
+      left,
+      width,
+      height,
+      zIndex: 1
+    };
+    element = (
+      <>
+        {!overscrolled && (
+          <ScrollerCell {...props}>
+            {children}
+          </ScrollerCell>
+        )}
+        <ScrollerCell {...props} style={style}>
+          {children}
+        </ScrollerCell>
+      </>
+    );
+  } else {
+    element = (
+      <ScrollerCell {...props}>
+        {children}
+      </ScrollerCell>
+    );
+  }
+
+  return element;
+
 };
 
 export default SpreadsheetCell;
