+++ src/components/Spreadsheet/useSpreadsheetRender.js
@@ -1,67 +1,4 @@
 import { useMemo } from 'react';
-import { getMergedCellSize, getMergedCellPosition } from './utils';
-
-const getMergedCellProps = ({
-  mergedRange,
-  rows,
-  columns,
-  defaultRowHeight,
-  defaultColumnWidth,
-  fixRows,
-  fixColumns
-}) => {
-  const columnIndex = mergedRange.start.column;
-  const rowIndex = mergedRange.start.row;
-  
-  const top = getMergedCellPosition({
-    meta: rows,
-    index: rowIndex,
-    defaultSize: defaultRowHeight
-  });
-  const left = getMergedCellPosition({
-    meta: columns,
-    index: columnIndex,
-    defaultSize: defaultColumnWidth
-  });
-
-  const width = getMergedCellSize({
-    count: mergedRange.end.column - mergedRange.start.column,
-    meta: columns,
-    startIndex: columnIndex,
-    defaultSize: defaultColumnWidth
-  });
-  const height = getMergedCellSize({
-    count: mergedRange.end.row - mergedRange.start.row,
-    meta: rows,
-    startIndex: rowIndex,
-    defaultSize: defaultRowHeight
-  });
-
-  const fixWidth = fixColumns && columnIndex <= fixColumns && getMergedCellSize({
-    count: fixColumns - mergedRange.start.column,
-    meta: columns,
-    startIndex: columnIndex,
-    defaultSize: defaultColumnWidth
-  });
-
-  const fixHeight = fixRows && rowIndex <= fixRows && getMergedCellSize({
-    count: fixRows - mergedRange.start.row,
-    meta: rows,
-    startIndex: rowIndex,
-    defaultSize: defaultRowHeight
-  });
-
-  return {
-    rowIndex,
-    columnIndex,
-    top,
-    left,
-    width,
-    height,
-    fixWidth,
-    fixHeight
-  }
-};
 
 /**
  * @param {import('./').useSpreadsheetRenderOptions} options
@@ -79,9 +16,7 @@ const useSpreadsheetRender = ({
   renderCellValue,
   fixRows,
   fixColumns,
-  mergedCells,
-  defaultColumnWidth,
-  defaultRowHeight
+  mergedCells
 }) => {
   const cellsElements = useMemo(() => {
     return visibleRows.reduce((acc, rowIndex) => {
@@ -117,9 +52,6 @@ const useSpreadsheetRender = ({
               return (mergedRange.start.row) === rowIndex &&
                   (mergedRange.start.column) === columnIndex
             });
-            const isFixedColumnArea = columnIndex <= fixColumns;
-            const isFixedRowArea = rowIndex <= fixRows;
-            const shouldRenderMerged = (isFixedColumnArea || isFixedRowArea) && mergedRange;
             
             let element;
 
@@ -129,7 +61,7 @@ const useSpreadsheetRender = ({
                 element = renderRowNumber({ row, column, rowIndex, columnIndex });
                 break;
               default:
-                element = renderCellValue({ row, rowIndex, column, columnIndex, value: curValue, mergedRange });
+                element = renderCellValue({ row, rowIndex, column, columnIndex, rows, columns, value: curValue, mergedRange });
                 break;
             }
 
@@ -151,72 +83,47 @@ const useSpreadsheetRender = ({
     renderColumnNumber,
     renderRowNumber,
     renderCellValue,
-    mergedCells,
-    fixRows,
-    fixColumns
+    mergedCells
   ]);
 
-  // Overscrolled cells. When merge starts out of visible cells area
+  // Overscrolled merged cells. When merge starts out of visible cells area
   const mergedCellsElements = useMemo(() => {
     // Filtering out merged ranges which are not visible
-    const mergedCellIsInRange = ({ meta, start, end, fixCount }) => {
-      return meta.indexOf(start) !== -1 ||
-          meta.indexOf(end) !== -1 ||
-          (start < meta[fixCount] && end > meta[meta.length - 1]);
+    const rangeIsOverscrolled = ({ meta, start, end, fixCount }) => {
+      return start < meta[fixCount] && end >= meta[fixCount];
     };
 
     return mergedCells ? mergedCells.filter(mergedRange => {
-      return mergedCellIsInRange({
+      return rangeIsOverscrolled({
             meta: visibleRows,
             start: mergedRange.start.row,
             end: mergedRange.end.row,
             fixCount: fixRows
           }) &&
-          mergedCellIsInRange({
+          rangeIsOverscrolled({
             meta: visibleColumns,
             start: mergedRange.start.column,
             end: mergedRange.end.column,
             fixCount: fixColumns
           })
     }).map(mergedRange => {
-      const {
-        rowIndex,
-        columnIndex,
-        top,
-        left,
-        width,
-        height
-      } = getMergedCellProps({
-        mergedRange,
-        rows,
-        columns,
-        defaultRowHeight,
-        defaultColumnWidth
-      });
-
-      const style = {
-        position: 'absolute',
-        top,
-        left,
-        width,
-        height
-      };
-
+      const columnIndex = mergedRange.start.column;
+      const rowIndex = mergedRange.start.row;
       const rowValue = value[rowIndex];
       const curValue = rowValue && rowValue[columnIndex];
-
       return renderCellValue({
-        columnIndex,
         rowIndex,
+        columnIndex,
+        rows,
+        columns,
         value: curValue,
-        style
+        mergedRange,
+        overscrolled: true
       });
     }, []) : [];
   }, [
     rows,
     columns,
-    defaultColumnWidth,
-    defaultRowHeight,
     fixColumns,
     fixRows,
     visibleRows,
